# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Federated Learning API",
        "description": "API documentation for the Medical AI Federated Learning platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/org-manager/members": {
            "get": {
                "tags": [
                    "OrgManager — Members"
                ],
                "summary": "List all members of this organization",
                "operationId": "592bfd4ed908830e6944f5bf6cf624d3",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/org-manager/members/{id}": {
            "get": {
                "tags": [
                    "OrgManager — Members"
                ],
                "summary": "Show a single member",
                "operationId": "3ef21b048b6238aa490af15353c95bf7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "OrgManager — Members"
                ],
                "summary": "Remove a member from the organization",
                "operationId": "dc1ef70a708642cbfbc61f858221675a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member removed"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Cannot remove yourself"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/org-manager/members/{id}/approve": {
            "post": {
                "tags": [
                    "OrgManager — Members"
                ],
                "summary": "Approve (activate) a pending doctor in this organization",
                "operationId": "8bab60ced74fddabc8af2781bcd37748",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member approved"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Member already active"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/org-manager/members/{id}/deactivate": {
            "post": {
                "tags": [
                    "OrgManager — Members"
                ],
                "summary": "Deactivate a member (suspend their access to this org)",
                "operationId": "69dc3c56e2a8cd4453d026ae9a4ab0d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member deactivated"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Cannot deactivate yourself"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/org-manager/patients": {
            "get": {
                "tags": [
                    "OrgManager — Patients"
                ],
                "summary": "List all patients in this organization",
                "operationId": "3ed3fdd4f979f6b32c109496031cd512",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "her2",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "er_status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of patients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PatientObject"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/org-manager/patients/{id}": {
            "get": {
                "tags": [
                    "OrgManager — Patients"
                ],
                "summary": "Show a single patient with their examination history",
                "operationId": "11d44a6a2ac00d50d1676741a45c4dff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PatientObject"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dr. Amina Belkacem"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "amina@clinic.dz"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+213661234567",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "organization": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 10
                            },
                            "name": {
                                "type": "string",
                                "example": "Ibn Sina Clinic"
                            },
                            "type": {
                                "type": "string",
                                "example": "clinic"
                            },
                            "status": {
                                "type": "string",
                                "example": "active"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "doctor"
                        ]
                    }
                },
                "type": "object"
            },
...
            "PatientObject": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "organization_id": {
                        "type": "integer"
                    },
                    "patient_identifier": {
                        "type": "string"
                    },
                    "er_status": {
                        "type": "boolean"
                    },
                    "pr_status": {
                        "type": "boolean"
                    },
                    "her2_binary": {
                        "type": "boolean"
                    },
                    "age": {
                        "type": "integer"
                    },
                    "stage_num": {
                        "type": "integer"
                    },
                    "er_status_missing": {
                        "type": "boolean"
                    },
                    "pr_status_missing": {
                        "type": "boolean"
                    },
                    "fraction_genome_altered": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "buffa_hypoxia_score": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "ragnum_hypoxia_score": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "winter_hypoxia_score": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "HttpOnly cookie issued by Sanctum after a successful `/api/verify-otp` call.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}